refactor(locks): centralize file
